[ADD] : SigUp FragmentStateAdapter
2022-05-27
