+++ presentation/src/main/res/layout/activity_sign_up.xml
@@ -38,10 +38,11 @@ xmlns:tools="http://schemas.android.com/tools">
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
             <androidx.viewpager2.widget.ViewPager2
-                android:layout_width="330dp"
+                android:layout_width="match_parent"
                 android:layout_height="500dp"
                 app:layout_constraintTop_toBottomOf="@id/image_sign_up"
-                app:layout_constraintStart_toStartOf="parent"/>
+                app:layout_constraintStart_toStartOf="parent"
+                android:id="@+id/view_pager"/>
             <com.google.android.material.button.MaterialButton
                 android:id="@+id/btn_sign_up"
                 android:layout_width="350dp"
@@ -56,7 +57,7 @@ xmlns:tools="http://schemas.android.com/tools">
                 android:textStyle="bold"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/layout_email" />
+                app:layout_constraintTop_toBottomOf="@+id/view_pager" />
 
             <androidx.appcompat.widget.LinearLayoutCompat
                 android:layout_width="match_parent"
