+++ presentation/src/main/res/layout/item_sign_up_2.xml
@@ -19,9 +19,9 @@
             style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
             android:layout_width="350dp"
             android:layout_height="wrap_content"
-            android:layout_marginStart="30dp"
+            android:layout_marginStart="15dp"
             android:layout_marginTop="30dp"
-            android:layout_marginEnd="1dp"
+            android:layout_marginEnd="15dp"
             app:boxStrokeErrorColor="#FF0000"
             app:counterEnabled="true"
             app:counterMaxLength="15"
